Fixed item icons from overlapping within the Options UI.
* Fix for overlapping icons in Manual Filter configuration

Setting the width to 'full' instead of leaving it unspecified avoids triggering a bug in AceConfig where the height of a widget is used before the widget width is resized causing the height to be recalculated.

* Fix for overlapping icons in Junk Filter configuration

Applied width = 'full' to the ItemLists in the junk.lua AceConfig options table. This works around a bug with widgets that use flow layout and resize as elements are added/removed or as the widget width is resized.