Refactor UpdateButton
Refactors UpdateButton to split Item logic from the Display logic. This was necessary due to multiple changes over the years that have added up and the initial assumption that only items are in your bags.  Things that arent items inlcude Battlepet Cages added in Mists of Pandaria and Mythic Keystones added in Legion.