Cache the Itemlink and only update when it changes
UpdateButton can be triggered numerous times even when the item for that button hasn't changed.  For example equipping an item will fire an UpdateButton for every button even if only 1 button needs to be updated.  This changed greatly reduces the number of times that updates are processed when not needed.