Fix: Section SetPoint Layout Bug
This corrects the `SetPoint failed` issue, which can happen
intermittently by resetting the `SetPoint` of each section at the
beginning of the layout.

As best I can tell, this is caused by the sort order changing between
layouts, which can cause occasionally cause any number of sections to
form a circular dependency.

Addresses #551