Resolve issues with container apis
The container APIs on 10.0.2 have moved into C_Container and a few functions
have been outright removed.

In addition the returns for GetContainerItemInfo and GetContainerItemQuestInfo
have changed to use tables instead of a long list of returns.

This introduces a set of wrapper functions to handle working with both the
new C_Container APIs and the legacy APIs to support wrath and classic-era.