fix(CurrencyFrame): tracking & options
Which currencies to show on the bagframe is now read from blizzards API directly. This means you can shift-click currencies in the "Currency list" panel and they will show up on the bagframe.

Because of this the copying of currencies to display in the plugin options is no longer needed and as such the options of each currency is no longer needed in the either.

Extended the options with a "spacing" option to control how much spacing there is between each currency tracked.